+++ peripherals/eic.h
@@ -60,27 +60,27 @@ void eic_enable(void);
   * @note Be sure to check your pin multiplexing table to ensure that you do not have multiple pins
   *       assigned to the same interrupt channel. Also note that the NMI pin is not currently supported.
   */
-int8_t eic_configure_pin(const uint32_t pin, eic_interrupt_trigger trigger);
+int8_t eic_configure_pin(const uint16_t pin, eic_interrupt_trigger trigger);
 
 /** @brief Enables an interrupt on the given interrupt channel.
   * @param pin The external interrupt pin.
   */
-bool eic_enable_interrupt(const uint8_t pin);
+bool eic_enable_interrupt(const uint16_t pin);
 
 /** @brief Disables the interrupt on the given interrupt channel.
   * @param pin The external interrupt pin.
   */
-bool eic_disable_interrupt(const uint8_t pin);
+bool eic_disable_interrupt(const uint16_t pin);
 
 /** @brief Enables an interrupt on the given interrupt channel.
   * @param pin The external interrupt pin.
   */
-bool eic_enable_event(const uint8_t pin);
+bool eic_enable_event(const uint16_t pin);
 
 /** @brief Disables the interrupt on the given interrupt channel.
   * @param pin The external interrupt pin.
   */
-bool eic_disable_event(const uint8_t pin);
+bool eic_disable_event(const uint16_t pin);
 
 /** @brief Configures an external interrupt callback.
   * @details You only get the one callback, so if you need to handle multiple interrupts, you'll need
